# backend/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/documents/{id}/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all versions newest first, each with fresh Cloudinary URLs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "List versions of a document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/documentversion.VersionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid document ID",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "List all versions (admin only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/documentversion.VersionResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — admin only",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/versions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "Get a version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/documentversion.VersionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Version not found",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "Delete a version (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/documentversion.VersionMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — admin only",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/versions/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns JSON with url + filename. The frontend triggers the download.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "Download a version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/documentversion.VersionDownloadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Version not found",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/versions/{id}/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Makes the given version the current one.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-versions"
                ],
                "summary": "Restore an old version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Version UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/documentversion.VersionRestoreResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Version not found",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "auth.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "invalid email or password"
                }
            }
        },
...
        "documentversion.VersionDownloadResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "Contract-2026.pdf"
                },
                "url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/..."
                }
            }
        },
        "documentversion.VersionMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Version deleted"
                }
            }
        },
        "documentversion.VersionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-09-21T14:48:32Z"
                },
                "document_id": {
                    "type": "integer",
                    "example": 12
                },
                "hashed_string": {
                    "type": "string",
                    "example": "a1b2c3..."
                },
                "id": {
                    "type": "string",
                    "example": "2f0c8e5b-..."
                },
                "original_filename": {
                    "type": "string",
                    "example": "Contract-2026.pdf"
                },
                "preview_url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/..."
                },
                "url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/..."
                },
                "user_id": {
                    "type": "integer",
                    "example": 3
                },
                "version": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
        "documentversion.VersionRestoreResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Version restored"
                },
                "version": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
    }
}